# Using SearchValues in ContentDispositionHeaderValue

Task of https://github.com/dotnet/aspnetcore/issues/46484

## Description

Updating the current Encode5987 method to use SearchValues and IndexOfAny method during the encode of the FileNameStar property. Adding unit tests and a benchmark to measure the before-after performance.

Compared to the .NET runtime [implementation](https://source.dot.net/#System.Net.Http/System/Net/Http/Headers/HeaderUtilities.cs,696f2b0de2ccd5de) a second `Span<char>` is pooled because `ValueStringBuilder` used by the .NET runtime is internal, hence `Encoding.ASCII.GetChars` needs a way to reference the chars passed to the `StringBuilder`

## Performance Comparison:

BenchmarkDotNet=v0.13.0, OS=Windows 10.0.22631
12th Gen Intel Core i7-1255U, 1 CPU, 12 logical and 10 physical cores
.NET SDK=9.0.100-preview.4.24201.1
  [Host]     : .NET 9.0.0 (9.0.24.20403), X64 RyuJIT
  Job-SUANLL : .NET 9.0.0 (9.0.24.18101), X64 RyuJIT

Server=True  Toolchain=.NET Core 9.0  RunStrategy=Throughput

BEFORE:

|                 Method |     Mean |   Error |  StdDev |        Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------- |---------:|--------:|--------:|------------:|-------:|------:|------:|----------:|
|   FileNameStarEncoding | 213.8 ns | 1.79 ns | 1.68 ns | 4,676,383.4 | 0.0052 |     - |     - |     496 B |
| FileNameStarNoEncoding | 201.0 ns | 1.32 ns | 1.23 ns | 4,975,577.6 | 0.0050 |     - |     - |     464 B |

AFTER **UPDATED: 2024/04/13 corresponding to commit [With custom utf8 and hex encoding combined](https://github.com/dotnet/aspnetcore/pull/55039/commits/2eadb693477f2556fc50787575bc325b3b22c6de)**:

|                 Method |      Mean |    Error |   StdDev |         Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------- |----------:|---------:|---------:|-------------:|-------:|------:|------:|----------:|
|   FileNameStarEncoding | 213.61 ns | 1.828 ns | 1.710 ns |  4,681,410.7 | 0.0052 |     - |     - |     496 B |
| FileNameStarNoEncoding |  87.12 ns | 0.962 ns | 0.853 ns | 11,478,216.6 | 0.0039 |     - |     - |     368 B |


Fixes #46484