@amcasey 2024-04-09
To confirm, `FileNameStarEncoding` is _worse_ after the change?