@ladeak 2024-04-09
> To confirm, `FileNameStarEncoding` is _worse_ after the change?

`Encoding` means it has characters to actually convert to HexString while `NoEncoding` does not have, but it still does make sure nothing to encode.
The actual results will also greatly depend on the number and frequency of chars to be converted to Hex.