@amcasey 2024-04-09
> > To confirm, `FileNameStarEncoding` is _worse_ after the change?
> 
> `Encoding` means it has characters to actually convert to HexString while `NoEncoding` does not have, but it still does make sure nothing to encode. The actual results will also greatly depend on the number and frequency of chars to be converted to Hex.

Sorry, I think we got our lines crossed.  I know encoding is expected to be slower than no-encoding (by a variable amount), but it looks like encoding-before is slower than encoding-after?