@ladeak 2024-04-10
Here is my latest run:

|                 Method |      Mean |    Error |   StdDev |         Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------- |----------:|---------:|---------:|-------------:|-------:|------:|------:|----------:|
|   FileNameStarEncoding | 211.84 ns | 1.965 ns | 1.838 ns |  4,720,566.7 | 0.0052 |     - |     - |     496 B |
| FileNameStarNoEncoding |  97.99 ns | 0.685 ns | 0.608 ns | 10,205,254.5 | 0.0039 |     - |     - |     368 B |