@ladeak 2024-04-12
With for loop to process the hex strings and using string interpolated overload of the string builder, I get the following perf results:


|                 Method |     Mean |   Error |  StdDev |        Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------- |---------:|--------:|--------:|------------:|-------:|------:|------:|----------:|
|   FileNameStarEncoding | 235.0 ns | 3.15 ns | 2.94 ns | 4,256,174.6 | 0.0052 |     - |     - |     496 B |
| FileNameStarNoEncoding | 101.3 ns | 1.75 ns | 1.63 ns | 9,867,874.7 | 0.0039 |     - |     - |     368 B |